redux-form: Improve FieldArray type definition
This change makes it much easier to use FieldArrays, there is no longer
a need to define a custom FieldArray type matching your data types.
The following was necessary before the changes in this commit:

    const CustomFieldArray = FieldArray as
        new () => GenericFieldArray<MyFieldValue, MyFieldArrayCustom...
    <CustomFieldArray component={ MyCustomArrayField } ... />

This commit renders the additional cast unnecessary and reduces the
above to the following:

    <FieldArray component={ MyCustomArrayField } ... />

Unfortunately, this new definition only works with typescript 2.7 and
above. The minimum typescript version was thus updated in index.d.ts.